# packages/server/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler.ownerResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
    type: object
  internal_handler.productRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      free_shipping:
        type: boolean
      image_url:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      name:
        type: string
      price_cents:
        type: integer
      stock:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price_cents
    type: object
  internal_handler.productResponse:
    properties:
      campaign_ends_at:
        type: string
      campaign_id:
        type: integer
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      effective_price_cents:
        type: integer
      free_shipping:
        type: boolean
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/internal_handler.ownerResponse'
      price_cents:
        description: |-
          PriceCents is the base price; EffectivePriceCents is what a buyer pays
          right now (equal to PriceCents when no campaign is live).
        type: integer
      slug:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: E-commerce backend API for kommers.
  title: Kommers API
  version: "1.0"
paths:
...
  /api/v1/products:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Filter by category
        in: query
        name: category_id
        type: integer
      - description: Search by product name
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handler.productResponse'
            type: array
      summary: List active products
      tags:
      - products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.productRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.productResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a product (admin or seller — becomes the owner)
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product (admin, or the owning seller)
      tags:
      - products
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.productRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.productResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a product (admin, or the owning seller)
      tags:
      - products
  /api/v1/products/{slug}:
    get:
      parameters:
      - description: Product slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.productResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an active product by slug
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"